cpp: libexec: Rename glview to view and showinf to info
This makes the commands match the documentation and top-level
wrapper.  They are still invocable as glview and showinf.
