+++ cpp/libexec/view/Window.h
@@ -36,25 +36,86 @@
  * #L%
  */
 
-#include <QApplication>
-#include <QDesktopWidget>
+#ifndef WINDOW_H
+#define WINDOW_H
 
-#include <glview/Window.h>
+#include <QMainWindow>
 
-using namespace glview;
+#include <ome/qtwidgets/GLView2D.h>
+#include <ome/qtwidgets/NavigationDock2D.h>
 
-int main(int argc, char *argv[])
+QT_BEGIN_NAMESPACE
+class QSlider;
+class QMenu;
+class QAction;
+class QActionGroup;
+QT_END_NAMESPACE
+
+namespace view
 {
-  QApplication app(argc, argv);
-  Window window;
-  window.resize(600,600);
 
-  float desktopArea = static_cast<float>(QApplication::desktop()->width() * QApplication::desktop()->height());
-  float widgetArea = static_cast<float>(window.width() * window.height());
-  if ((widgetArea / desktopArea) < 0.75f)
-    window.show();
-  else
-    window.showMaximized();
+  class Window : public QMainWindow
+  {
+    Q_OBJECT
 
-  return app.exec();
+  public:
+    Window();
+
+  private slots:
+    void open();
+    void open(const QString& file);
+    void quit();
+    void view_reset();
+    void view_zoom();
+    void view_pan();
+    void view_rotate();
+    void viewFocusChanged(ome::qtwidgets::GLView2D *glView);
+    void tabChanged(int index);
+
+  private:
+    void createActions();
+    void createMenus();
+    void createToolbars();
+    void createDockWindows();
+
+    QMenu *fileMenu;
+    QMenu *viewMenu;
+
+    QToolBar *Cam2DTools;
+
+    QAction *openAction;
+    QAction *quitAction;
+
+    QAction *viewResetAction;
+
+    QActionGroup *viewActionGroup;
+    QAction *viewZoomAction;
+    QAction *viewPanAction;
+    QAction *viewRotateAction;
+
+    QSlider *createAngleSlider();
+    QSlider *createRangeSlider();
+    ome::qtwidgets::NavigationDock2D *navigation;
+
+    QTabWidget *tabs;
+    ome::qtwidgets::GLView2D *glView;
+    QSlider *minSlider;
+    QSlider *maxSlider;
+
+    QMetaObject::Connection minSliderChanged;
+    QMetaObject::Connection minSliderUpdate;
+    QMetaObject::Connection maxSliderChanged;
+    QMetaObject::Connection maxSliderUpdate;
+    QMetaObject::Connection navigationChanged;
+    QMetaObject::Connection navigationUpdate;
+  };
+
 }
+
+#endif // WINDOW_H
+
+/*
+ * Local Variables:
+ * mode:C++
+ * End:
+ */
